feat(policies): add pundit policies to needed actions

+++ app/controllers/child_presents_controller.rb
@@ -15,6 +15,8 @@ class ChildPresentsController < ApplicationController
   end
 
   def destroy
+    authorize ChildPresent
+
     @present = @child_profile.child_presents.find(params[:id])
 
     if @present.destroy
